+++ Common/DtaHashPwd.cpp
@@ -46,7 +46,7 @@ void DtaHashPassword(vector<uint8_t> &hash, char * password, vector<uint8_t> sal
 		goto exit;
 	}
 	hash.reserve(hashsize + 2); // hope this will prevent reallocation
-	for (uint16_t i = 0; i < hashsize; i++) {
+	for (size_t i = 0; i < hashsize; i++) {
 		hash.push_back(' ');
 	}
 	
@@ -70,7 +70,7 @@ void DtaHashPwd(vector<uint8_t> &hash, char * password, DtaDev * d)
 
     if (d->no_hash_passwords) {
         hash.clear();
-	for (uint16_t i = 0; i < strnlen(password, 32); i++)
+	for (size_t i = 0; i < strnlen(password, 32); i++)
 		hash.push_back(password[i]);
 	// add the token overhead
 	hash.insert(hash.begin(), (uint8_t)hash.size());
@@ -115,7 +115,7 @@ int Testsedutil(const PBKDF_TestTuple *testSet, unsigned int testSetSize)
         const PBKDF_TestTuple &tuple = testSet[i];
         hash.clear();
         seaSalt.clear();
-        for (uint16_t j = 0; j < strnlen(tuple.Salt, 255); j++) {
+        for (size_t j = 0; j < strnlen(tuple.Salt, 255); j++) {
             seaSalt.push_back(tuple.Salt[j]);
         }
 		printf("Password %s Salt %s Iterations %i Length %i\n", (char *)tuple.Password,
